requirements.txt
discord.py
websocket-client
requests
coloredlogs